Correct behavior of move or merge
archivematica/issues#1018 documents that Archivematica previously
assumed a `metadata` folder existed, or would be created in the
correct location in a potential AIP during the ingest workflow.

This commit fixes a regression that caused this directory not to be
created when a 'src' location was otherwise empty. If a top-level
directory doesn't exist it will now get created and the 'src'
will be moved entirely rather than being copied one-by-one.