Re-enable package.py to update its file list
Before c1e7468 Archivematica package objects were able to be updated
with new file information. c1e7468 removed this asking the files
associated with a package to be returned from either the database
when objects were known to be in the database, or the file system
when no database objects existed.

Archivematica currently relies on being able to update a package
object's associated files with those on disk as well as in the
database where microservices are interacting e.g. creating or
updating file information in contrasting and potentially
inconsistent ways.

Tests simulate the interplay between the filesystem and database
when there are objects in both locations that need to be looked
after.

Co-authored-by: Douglas Cerna (Soy Douglas) <douglascerna@yahoo.com>