Identify workflow endpoints
This commit ensures that workflow endpoints are identified that when
reached should reduce the number of active packages in the
Archivematica workflow. This becomes critical when Archivematica
only recognizes one core available to be used on a machine. If the
count is 1 and the cores are 1 then no new packages can be generated.
We need to increment and decrement our internal counters as reliably
as possible.