Reorder processing configuration choices (#1502)
* WIP: Reorder processing configuration choices

I've gone through each of the processing configuration choices and
reordered where necessary. Many of the changes are just for consistency
- I've changed the order to `None`, `Yes`, `No` where those were the
only options. For some of the jobs with more complex options I've tried
to apply a more logical order:

```
Normalize:
* Normalize for preservation and access
* Normalize for preservation
* Normalize for access
* Normalize service files for access
* Normalize manually
* Do not normalize
```

```
Select compression level:
* 1 - fastest compression
* 3 - fast compression
* 5 - normal compression
* 7 - maximum compression
* 9 - ultra compression
```

```
Upload DIP:
* Upload DIP to AtoM/Binder
* Upload DIP to ArchivesSpace
* Upload DIP to CONTENTdm
* Do not upload DIP
```

One other change I've made here is in the `Generate thumbnails` option,
where I've clarified the options:

```
Generate thumbnails:
* Yes
* Yes, without default icons
* No
```

The reason that this is still a WIP is because while the changes appear
on the processing configuration page (Administration > Processing
configuration > Default), they're not all reflected on the
Transfer/Ingest tab drop-down menus. I'm not sure if this is a
not-knowing-enough-about-docker problem or if there's something else in
the code that I need to change.

Connected to #891

* Change "Delete packages after extraction" options

Changed for consistency into `None`, `Yes`, `No` order.

* Tidy up indenting

* Make _pull_choices respect workflow ordering

With @replaceafill's help, I've made a couple small changes to force
the `_pull_choices` helper to respect the ordering of workflow choices
in workflow.json.

* Reorders final few items

* Formatting fix

* Set choices as OrderedDict

Co-authored-by: Ashley <ashley.blewer@gmail.com>