Reorder reingest dropdown options (#1568)
* Reorder reingest dropdown options

This was missed in the original PR, since I was using the processing
config for guidance. I've just switched the order of these two options
so that "Approve" is first, followed by "Reject", to be consistent with
all of the other dropdowns.

Connected to archivematica/Issues#819

* Change order of Approve SIP creation dropdown

This is the dropdown you get when you move something from backlog into
Ingest - microservice `Approve SIP creation`, `Job: Approve SIP
Creation`.