Undo PR #1561
In https://github.com/artefactual/archivematica/pull/1561 we removed
the creation of the base directory for a Dspace transfer from the
`fileToFolder_v1.0` client script because the MCPServer created the
directory in advance.

Since we're removing that capability from the MCPServer in this PR, we
need to restore the old logic in the client script back.