check_transer_dir_for_objects: set error path
When the transfer directory does not contain any objects we need to fail
the transfer. In case of a non zero exit code we need to fall back to
the "Email fail report" step. This step is also configured as the
default fallback in case anything else goes wrong.

This commit includes two small changes to the
check_transfer_dir_for_objects script. First of all it changes the exit
codes to more intuitive scheme. In case there are objects the exit code
is 0 (success) and the frontend will give a green exit code sign. In
case of no objects the script returns 1 and the frontend will display a
red exit code sign.

This commit also simplifies the if check for existing files. Files is a
is a list of the names of the non-directory files in the directory path.
This means that files is either an empty or non-empty list, and maybe
but not likely, None. In case the list is empty or None, there are no
files, otherwhise there are files. This behaviour is nicely captured by
the default boolean resolution of the Python list:

In [1]: files = []
In [2]: bool(files)
Out[2]: False

In [3]: files = ['foo', 'bar']
In [4]: bool(files)
Out[4]: True

In [5]: files = None
In [6]: bool(files)
Out[6]: False

This commit also transforms some variables in the script to snake_case
to match default Python code style and the code style used in most of
the Archivematica project.