workflow: add is_terminal check
Add an is_terminal property to the workflow. With this property it is
possible to indicate for every link if it is the end of a workflow. This
can be used by the queueing mechanism to see if the next job in the
chain needs to be scheduled, if a package needs to be (de)activated, or
if a new job needs to be queued from one of the existing queues.

Using this property is as simple as adding the boolean field "end" to
any link in the existing workflow. Using the property is as simple as
calling is_terminal() on any Link. This will retrun the value of the
"end" field, or false if the field does not exist.