queues: use workflow to indicate package ending
Set the "end" property on workflow links that indicate the end of the
processing of a package. Use this property to decide if a package needs
to be deactivated. This allows for manipulating the queueing and package
behaviour from the workflow, instead of hardcoded in the queueing logic.

In the future we might want to switch the is_terminal property to a less
explicit check. Instead of checking for the "end" property it should
check if there are no links to future jobs. If this is the case this
would signal the end of the chain. See issue #1055 for the full
discussion on this topic.

This commit also clarifies the behaviour of the package and job complete
callbacks. It specifically mentions the role of the callbacks with
respect to the job chains.