queues: limit number of queued packages
This commit adds a configurable limit to the maximum number of packages
that can be waiting to be processed. In the old situation this number
was set to 4096. This means that any automated system (or very active
user) could schedule transfers until a disk size limits is hit.

The default, 128, is still high this would make sure that the most users
won't see any changes. However, users using the API will get errors when
they set a more sensible default and submit to much packages.

The system is implemented by using a bounded transfer queue. If too many
transfers are created and scheduled, schedule_job will run into a
Queue.Full exception at some point. To prevent this from happening we
introduce the concept of transfer slots. Before queueing a transfer, you
can request a transfer slot. If the transfer slot is granted you can
proceed to schedule the job, otherwhise the transfer should not be
started.

To make sure that every transfer that is queued has requested a slot
before queuing we have chosen a bounded semaphore. This semaphore will
make sure that it can never be released more often than it is acquired.
Each time we remove a transfer from the queue we release the semaphore,
if we remove more transfers than we have acquired slots we log an error.

fixup! queues: limit number of queued packages