packages: acquire transfer slot
The queueing mechanism poses limits on the number of transfers that can
be queued at any given moment. To make sure that this limit is not
reached we can request a transfer slot. This commit adds the transfer
slot capabilities to the packages module.

There are two changes. The first change is that before creating a
transfer a transfer slot is aqcuired. The second change is a bit more
subtle. Transfer slots are automatically released when a transfer is
started, however if a transfer is never due to an error the slot is
lost. To keep this from happening we have added a callback that releases
the transfer slot in case of problems during the creation of a transfer.