dashboard api: expose errors on package endpoint
It should be possible for users to correct simple mistakes without
having to check the logs or having to consult a sysadmin. Besides this
the user should be able to differentiate between a bad API call and
hitting the transfer queue limit.

Currently the only two errors that are exposed are missing fields in
case of a ValueError and an error indicating that the user has hit the
transfer queue limit.