Proposal: Add v2beta/package validation in Dashboard
The API specification for this endpoint doesn't say anything about
the exact form of the JSON response, only that JSON will be
returned.

In this instance, now Transfer Types are more centralzed, we can
promote validation to the dasboard to provide greater feedback to
the end-user.

More importantly this validation helps to decrease the chances of
regressions in Archiveamatica by better supporting the unit testing
of transfer types.