MCPServer: add terminal links to workflow
* queues: update PACKAGE_COMPLETED_LINK_IDS

Add bbfbecde-370c-4e26-8087-cfa751e72e6a (Failed Compliance) as a
terminal link. If we don't we can block the system if too many packages
fail.

* workflow: add is_terminal check

Add an is_terminal property to the workflow. With this property it is
possible to indicate for every link if it is the end of a workflow. This
can be used by the queueing mechanism to see if the next job in the
chain needs to be scheduled, if a package needs to be (de)activated, or
if a new job needs to be queued from one of the existing queues.

Using this property is as simple as adding the boolean field "end" to
any link in the existing workflow. Using the property is as simple as
calling is_terminal() on any Link. This will retrun the value of the
"end" field, or false if the field does not exist.

* queues: use workflow to indicate package ending

Set the "end" property on workflow links that indicate the end of the
processing of a package. Use this property to decide if a package needs
to be deactivated. This allows for manipulating the queueing and package
behaviour from the workflow, instead of hardcoded in the queueing logic.

In the future we might want to switch the is_terminal property to a less
explicit check. Instead of checking for the "end" property it should
check if there are no links to future jobs. If this is the case this
would signal the end of the chain. See issue #1055 for the full
discussion on this topic.

This commit also clarifies the behaviour of the package and job complete
callbacks. It specifically mentions the role of the callbacks with
respect to the job chains.

* workflow: apply review comments

* queues: apply review comments

* queues: fix docstring comments

* workflow: add review comments

* Add 16415d2f because at this point a SIP is created for TRIM
  transfers.

* Add f8ee488b and 653b134f, these signal the end of the DIP packages.

* Exchange bbfbecde for 89071669, the package is done after the error
  path is completely finished. This is as soon as the transfer has
  reached the "normal" flow again.