MCPServer: relax logging on unexpected terminal links
Our workflow has a number of terminal links that may actually continue doing
work, e.g. in case of errors. Under these circumstances, MCPServer logs an
ERROR but it allows processing to continue assuming that termination happens
somewhere down the same chain.

E.g. "Move transfer to backlog" is one of the possible ends of a transfer
package, unless the operation fails in which case we've manually confirmed that
the error path terminates properly - this last step is important otherwise our
queues are exhausted blocking system processing generally.

The ERROR logging level is misleading for sysadmins since we don't expect them
to correct the situation. This commit lowers the level to WARNING.

amflow can help identify anomalies, e.g. these are "false positives" found with
`amflow check` at the time of writing this commit:

    [abd6d60c-d50f-4660-a189-ac1b34fafe85]
    Move transfer to backlog [children=1] [refsWD=false]

    [16415d2f-5642-496d-a46d-00028ef6eb0a]
    Create SIPs from TRIM transfer containers [children=2] [refsWD=true]

    [653b134f-4a37-4578-a286-7f2072e89f9e]
    Store DIP [children=1] [refsWD=false]

We've confirmed that the alternative paths take care of termination.

In the long term, it may be preferably to explore other ways to signal end of
processing.

Connects to https://github.com/archivematica/Issues/issues/1108.