Use get_model to get models imported from wsgi.py
In Django 1.9+ is not possible to import models before their
applications have been loaded.

The `elasticSearchFunctions` module (which imports File, Transfer and
Identifier) is used for initialization in the wsgi setup. The
`archivematicaFunctions` module (which imports DashboardSetting) is
imported indirectly too.

This changes the explicit import of the modules to use the `get_model`
helper in the application registry.