Add fields and column visibility to Archival Storage
THis work rewrites much of the Archival Storage tab to use the
DataTables library, adds new information such as accession number
and file count to the search results table, and adds a column
visibility selector that persists state to DashboardSettings.

The "Show AICs" checkbox has been removed, as it did not include
or exclude AICs from search results but rather determined whether
the "AIC" column was shown in the table. Since column visibility
is now the domain of the DataTable and new visibility selector, the
checkbox has been removed.

Additionally, the "Create an AIC button" is always displayed as
long as the "Show files?" checkbox is not checked. Previously, the
button would only appear after a user specifically ran a search from
the search bar at the top of the page.

Changes in the Elasticsearch mapping (specifically, adding the
filePath.raw field in the aipfiles index) will need to be applied
to existing ES indexes through reindexing or Put Mapping. Otherwise,
sorting on some columns may be broken.