Update commit message section of CONTRIBUTING.md
The "Commit messages" section of the CONTRIBUTING.md linked to a section
of the Archivematica docs style guide that has been removed.

I've changed the section to reference Chris Beams' "How to Write a Git
Commit Message" up front and included the seven rules from that blog
post directly, so that it's preserved even if the blog post disappears.

Connected to archivematica/Issues#1181