View fixes
In Django 1.11 the context for rendering templates from the backend
has to be a dictionary now. This commit removes the RequestContext
objects from several calls.

A TEMPLATES setting has also been added to the MCPClient because it's
needed by the normalization report and the pid binding scripts.

And required form fields now have the required HTML attribute by
default. This commit also updates a few instances where forms can't be
submitted because of this.