Add CSRF middleware
This adds Django's `CsrfViewMiddleware` to the dashboard settings and
the `{% csrf_token %}` tag to the form templates.

The wrapper for the API endpoints has been marked as CSRF exempt to
allow API calls to reach authentication, but the CSRF checks are still
performed by the Tastypie's authentication backends when needed.