Add management command to update ES mappings
This management command uses Put Mapping to update the mappings for the
aips and aipfiles indices, to ensure proper sorting in the Archival
Storage tab DataTable, and runs an Update By Query to populate the new
filePath.raw subfield in the aipfiles index.

This command must be run immediately following the upgrade to
Archivematica 1.12, prior to new AIPs being ingested, as Put Mapping is
unable to change the data type of existing ES fields. New string fields
will be automatically set to "text" rather than "keyword" if not
explicitly defined in the mapping, which will prevent the Put Mapping
command from working.