Turn off "show files" checkbox input autocomplete
This fixes a bug in Chrome where navigating from the "show files"
results in the Archival Storage tab to an AIP detail page and then
using the browser's back button caused a DataTables error.

This commit also removes a duplicative template import of misc.js,
which was discovered during debugging.