Enable the download of the AIP ES index as CSV
This commit enables the download of the AIP Elasticsearch index as a
CSV file, adding the all-important "Download CSV" button to the user
interface, and creating various helpers throughout the Javascript and
Python code. Cleaning up various components as we go, for example,
introducing more DRY variables to the Elasticsearch module following
the example of the DataTables enhancements. CSV handling is handled
by Pandas, because Pandas are awesome.

Pandas is a general purpose data handling library that provides a
nice way to bring data together, manipulate it as necessary, and
output it in multiple formats should we want them in future.
Importantly, it reduces the amount of handcrafting involved in
generating CSVs.

Existing Elasticsearch indices will not contain location, only newly
built indices. As such, the reindex scripts have also been updated
for those who wish to take advantage of this feature for past AIPs.

Tests have been added for the feature. Translation files are ready
to be translated!