Update ES following rejected deletion request
This commit fixes a bug described in issue #1273, whereby the statuses
for transfers and their files, as well as for AIP files, were not
updated after a package deletion request was rejected by a Storage
Service administrator. It also modifies the Backlog rendering function
to display "Deletion requested" rather than "Delete requested" for the
Status column, to fix an inconsistency between the Backlog and Archival
Storage tabs.