Replace METS structural component generation
flieSec and structMap generation is now done through components
cherry-picked from Cole's transfer METS creation module. The work
temporarily duplicates some of that effort to help us demonstrate
quicker, and then to help us find the commonalities. Which may be
many but for now, we really just need the fileSec and structMap.