Remove longerusername
longerusername increases the maximum length of Django's username field to 255
characters. That change seems to conflict with a migration added by Django
1.10 that increased the size from 33 to 150 characters (see link below to know
more).

This commit removes the longerusername app since Django solves the problem for
the most part and it does it in a way that is more sensible with MySQL
limitations. For users upgrading from an older release, we can advise to
manually change the schema - this is something we could probably do from a
migration but with the risk of trimming usernames when they exceed the limit.

https://docs.djangoproject.com/en/1.11/releases/1.10/#abstractuser-username-max-length-increased-to-150