Report size even if du returns non-zero exit code
When du encounters directories for which it doesn't have permissions it
returns an exit code of 1, causing subprocess.check_output to throw a
CalledProcessError. This resulted in `usage_get_directory_used_bytes`
returning 0 when run against a directory to which the archivematica
user did not have full permissions.

This commit modifies how a non-zero du exit code is handled. Now, a
warning is logged but the bytes reported by du are read from
CalledProcessError.output so that the total number of bytes in
directories the archivematica user can access is still reported in the
dashboard's Processing storage usage page.