Fix the use of the Dashboard's polling interval
This fix collects a number of smaller commits that work toward
improving how often the Dashboard communicates  with the
Archivematica back-end to request information.

We now use the jobs.js constructor to set Jobs polling interval.
The class constructor is called after the window is ready which means
that we can safely use it to set the polling interval where
previously the change in the dashboard to the Django setting
POLLING_INTERVAL would have no effect.

The modal dialog has been cleaned up. As this is a good opportunity
to do that while testing.

A previously unused application setting was removed. Per the original
ticket archivematica/issues#1017 and with a full search of the code,
this setting is unused and looks to be the remnants of a past
long-gone 🦕.

The polling interval default has been upped to 10 and can now be
configured in the system's environment. The README has been
updated to reflect this.