MCPServer: fix flaky tests
Since they were conceived, tests in test_queues randomly fail. In my tests,
that happens approximiately 1 every 100 times you run them. We've also seen
this in CI.

This commit blocks the thread momentarily before assertions are made so the
queue manager has time to transition into the desired state. Sometimes, this
seems to happen slower than the assertions are executed and that's what was
causing the tests to fail. Since we're working with threads, results are
non-deterministic.