Ensure get_directory_name doesn't return None
We had a situation where we were attempting to retrieve a transfer
name from a path member variable belonging to the Job class, but
the regular expression tests used to pull the transfer name from
the path could fall-through and return None. This resulted in an
empty transfer string in the Transfer/Ingest tab.

Here we add unit tests and ensure that a default is always returned
which will be the SIP UUID.