Add password validators
This commit adds password validation to the Archivematica dashboard. The
validators employed include:
- Django's MinimumLengthValidator (default length: 8 characters)
- Django's CommonPasswordValidator
- Django's UserAttributeSimilarityValidator
- A custom PasswordComplexityValidator, which checks that passwords contain at
least 3 of: lower-case characters, upper-case characters, numbers, special
characters (symbols or other non-alphanumeric Unicode characters).

The validators are enabled by default but can be disabled individually via the
environment. When enabled, validators are enforced when creating or editing
users via the web UI or the createsuperuser and changepassword management
commands.