Delete set_agent_code management command
set_agent_code is an undocumented management command that we have used
in the past when running forks of Archivematica where we wanted to use
a different preservation system identifier.

In Archivematica 1.13, the preservation system identifier is not going
to be maintained in the database but generated by the `version` module.
Forks can tweak that module as needed.

Connects to https://github.com/archivematica/Issues/issues/1336.