Generate and upload coverage report
This modifies the `test` CI workflow to generate coverage reports for
the `make test-...` rules of the MCPServer, MCPClient, dashboard and
archivematicaCommon.

These reports are then uploaded and processed in codecov.io.

The `codecov.yml` configuration is a copy from enduro with an extra
fix to map the internal /src/ path of the Docker containers.