Disallow hiding active packages
This commit makes the "active" attribute of a package available to the
front-end so it is posible to reject user requests attempting to hide
them.

A window.alert dialog is used to reject the user action, but this can
be later improved, e.g. hiding the button which needs some refactoring
in the rendering layer.

A similar check is added to the back-end endpoint. The `mark_hidden`
view now rejects hiding active packages returning a 409 conflict code.