Simplify dict normalization for ES indexing
Elasticsearch fields accept a single value or a list of values so there
is no need to make a distinction in that case, only between strings and
dictionaries. The previous normalization named dict fields always as
"tag_dict_list" even if there was only one value and it named string
fields as "tag" (for a single value) or "tag_str_list" (for multiple
values). The new normalization only differentiates between dicts
("tag_dict") and strings ("tag").