Do not overwrite PYTHON_VERSION build argument
If the environment variable PYTHON_VERSION is not set, the make rule
currently passes an empty argument to the Docker build which makes
pyenv to fail.

This fixes it by relying on the Docker's mechanism to ignore not
exported variables.

See https://docs.docker.com/engine/reference/commandline/run/#env