Fix accession event helper
The escape_string function has been removed in recent versions of
mysqlclient which makes the accession event helper to fail. Thanks to
Jesús for reporting it!

I don't think we need to use that mysqlclient function anymore. It
seems to come from days when we were doing raw SQL queries.