Fix validation of chain_choices and replacements
Only the first item in the `chain_choices` and `replacements` arrays was being
validated due to our usage of an array of schemas, leading to tuple validation.
In this context, each position in the array was expected to validate against
the schema at the corresponding position in the `items` array, leaving
subsequent items unvalidated.

This commit addresses this by converting `items` into a single schema object,
ensuring that validation applies to all items within the array, not just the
first.