Update MCPClient forking model
This introduces a new prefork execution model in MCPClient based on a
process pool.

The size of the pool is configurable and dynamically adjusted by the
parent process. Workers are automatically restarted to free resources
and will reuse database connections when possible.

Logging is centralized with a queue that is shared with pool
processes. Metrics are captured using mmap and collected by the parent
process.

Co-authored-by: Cole Maclean <hi@colemaclean.dev>
Co-authored-by: Jesús García Crespo <jesus@sevein.com>
Co-authored-by: Miguel Angel Medinilla <mamedin@gmail.com>