Upload AMAUAT logs based on outcome of the tag run
This generates and uploads the Docker logs of the AMAUAT workflow only
if the `Run AMAUAT tag` step either fails or gets cancelled.

This gets us the logs even if GitHub times out the workflow run when
the tag execution hangs or if the developer cancels it manually, and
avoids getting empty logs on Docker build failures.