Replace handle_unstored_dip client script
This removes the `handleunstoreddip_v0.0` client script used when a
DIP is rejected and replaces it with `move_v0.0`. This means the DIP
will always be rejected if it's not stored.

The workflow labels have been updated to be more similar to the AIP
workflow and the `DIP-` prefix has been added to the rejected DIP
directory to avoid name clashes with rejected AIPs.