Make uploadedDIPs a real watched directory
* Make uploadedDIPs a real watched directory

This adds a `watched_directories` definition for the directory and
chain handler that points to the `Store DIP` decision point.

* Move the uploaded DIP instead of copying it

This uses the `move_v0.0` client script instead of `copy_v0.0` to put
the DIP in the new chain. It also removes the link to the next job
which completes the running `uploadDIP` chain.

* Remove DIP directory after storing it

This removes logic for deleting watched directories from the store DIP
job. Instead, it adds a separate job to do so which makes DIP deletion
independent from the DIP upload process and from any interaction with
the space of the Storage Service.

* Replace handle_unstored_dip client script

This removes the `handleunstoreddip_v0.0` client script used when a
DIP is rejected and replaces it with `move_v0.0`. This means the DIP
will always be rejected if it's not stored.

The workflow labels have been updated to be more similar to the AIP
workflow and the `DIP-` prefix has been added to the rejected DIP
directory to avoid name clashes with rejected AIPs.

* Remove trailing slash from DIP path in Store DIP job

This client script assumes the DIP path doesn't end with a trailing slash,
but the mcp.watched_dir_handler of the MCPServer appends it to the
`%SIPDirectory%` replacement variable which populates the
`aip_filename` argument of the script.

* Update archivematica-acceptance-tests submodule