Replace clamd with clamav-client
This commit replaces [clamd 1.0.2] with [clamav-client 0.6.0], a modern fork of
clamd that follows current Python practices and tooling. clamav-client is tested
with supported versions of Python and avoids deprecated dependencies such as
`pkg_resources` (which was removed in Python 3.12).

clamav-client provides its own `Scanner` interface to handle integration with
both `clamd` and `clamscan`, so Archivematica doesn't need to. Both backends are
tested with integration tests.

[clamd 1.0.2]: https://pypi.org/project/clamd/1.0.2/
[clamav-client 0.6.0]: https://pypi.org/project/clamav-client/0.6.0/