Enhance OIDC authentication flow
This adds two new enhancements for OIDC authentication flow:

* Add a new setting to control whether local AM authentication is
available when OIDC authentication is in use. If local AM
authentication is disabled, then users will only be able to
authenticate via the OIDC provider. If the new setting is not
configured, local AM authentication is available.

* Add ability to define more than one OIDC provider in AM. Specific
providers can be chosen using HTTP query params passed to the server
when authenticating.