Add test case for empty bag metadata
This simulates the scenarios where a bag does not contain a
bag-info.txt and bagit returns an empty dictionary for its info
attribute or when reading the bag goes wrong and the create_mets_v2
script logs the problem but also returns an empty dictionary.

These scenario raise the same exception shown in issue #1129.