Set webpack dev mode to 'source-map'
Using the webpack default devtool mode causes CSP errors due to the use
of eval() statements in the pack file. CSP headers cannot be tested when
using 'eval-source-map' as it generates a large number of violations due
to the inline scripts and styles run using eval(). Setting 'source-map'
does not have this issue. If CSP is in use, 'eval-source-map' cannot be
used.